Add ktor module and make necessary changes for integration
This commit adds ktor module to the project and makes necessary changes to existing codebase for integration. The ktor module was added to expand the functionality of the project and use the benefits of web application framework ktor. Some of the changes include moving logic for document parser (verification, document loading), enhancing the Context class and adjustments for asynchronous communication. The utility functions for AWS S3 storage operations and some useful coroutines extension functions were also included. JWT authentication module was implemented as well. In addition to this, some entities related to user and some API endpoints were implemented. (This commit message was generated by JetBrains AI)