Refactor S3 upload methods and introduce MultipartExtractor
Refactored the S3 upload methods to have more descriptive names, improving readability. Deleted the Multipart.kt file as it had an implementation that raised exceptions and introduced potential for errors. Instead, a new class, MultipartExtractor, was introduced for safer and clearer processing of MultiPart data. This class includes a new utility method to simplify extracting multipart data from a call. The method names in the StudentDocumentService were also updated, following changes in S3Extensions. Adjustments were also made in StudentsController and Verifier for this change. The optional 'stage' attribute was added to Message dto for flexibility, and an unnecessary 'launch' import was removed from Application.kt for clean up.