Refactor application structure for different configurations
Modified the entire application structure to support different configurations for environment development and production. Broke down the previous single application.yaml into individual application-dev.yaml and application-production.yaml for specific configuration based on the environment. Also, changes made to Application.kt for loading the corresponding yaml based on the environment variable 'ktor.profile'. Removed the BCrypt.kt file and configuration as it was not required. Added logging support. Updated naming and application properties to meet the new structure requirements. Updated JWTService to receive jwtTokenExpiration from JWT config.

These changes improve the flexibility of the application configuration based on the environment it's deployed in, providing ease of management and increased security in production environments.