Refactor 'Context' to 'VerificationContext' and update imports
This commit refactors 'Context' to 'VerificationContext' in the application. This change was made as the context used throughout the application is more
specific to verification processes rather than general context. The renaming provides a more descriptive and accurate representation of its function.

In line with this, all imports and uses of 'Context' have also been updated to 'VerificationContext'. Some functions are also updated to be `suspend`
functions in order to fit with the refactored context model.

Also, a new testing library, "io.kotest:kotest-runner-junit5:5.8.0", is introduced in "core/build.gradle.kts" to replace "org.junit.jupiter:junit-jupiter:5.10.1".